tests/wpt/web-platform-tests/old-tests/submission/Infraware/Forms/contents/Forms/input_value_invalidstateerr.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>input_value_INVALID_STATE_ERR</h3>
  </p>
  
  <hr>

  <div id="log"></div>

  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action=""
      name="input_form">
  <p><input type='file' id='input_file'></p>
  </form>

  <script>
    
    var input_file = document.getElementById("input_file");
    try {
      input_file.value = "val";
      test(function() {
        assert_unreached("INVALID_STATE_ERR error is not raised.");
      });
    } catch (e) {
      test(function() {
        assert_equals(e.code, e["INVALID_STATE_ERR"], "INVALID_STATE_ERR error is not raised.");
      });
    }

  </script>

 </body>
</html>
